requirements.txt
numpy
setuptools
termcolor
absl-py
dotenv
pandas
requests
pytest
pytest-cov
python-dateutil
reactivex
retry
IPython
typing-extensions
jinja2
hf_xet
Pyarrow
sentence-transformers
# Concordia Assistant dependencies
PyYAML
chromadb


# Documentation generation
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints
myst-parser